moves printCombo to HTMLController, removes class GUI, replaces every usage of printCombo to point to its new location
