+++ src/xhtml/XHtmlSimpleElement.php
@@ -19,6 +19,7 @@ class XHtmlSimpleElement
     public $_htmlcode;
     public $_attributes = [];
 
+    use \PHPPgAdmin\HelperTrait;
     /**
      * Constructor
      *
@@ -45,8 +46,11 @@ class XHtmlSimpleElement
 
     public function is_element()
     {
-        return
-        str_replace('phppgadmin\xhtml\xhtml_', '', strtolower(get_class($this)));
+        $lower_classname   = strtolower(get_class($this));
+        $is_element_string = str_replace('phppgadmin\xhtml\xhtml_', '', $lower_classname);
+        //$this->prtrace('is_element_string', $is_element_string, 'lower_classname', $lower_classname, '__CLASS__');
+        return $is_element_string;
+
     }
 
     /**
@@ -55,14 +59,14 @@ class XHtmlSimpleElement
      */
     public function _html()
     {
-        $this->_htmlcode = "<";
+        $this->_htmlcode = '<';
         foreach ($this->_attributeCollection as $attribute => $value) {
             if (!empty($value)) {
                 $this->_htmlcode .= " {$attribute}=\"{$value}\"";
             }
 
         }
-        $this->_htmlcode .= "/>";
+        $this->_htmlcode .= '/>';
 
         return $this->_htmlcode;
     }
